chore: reorganize the response classes
Reorganize the response classes to reduce the number of top level files.

Combine builders and response classes.
