--- lib/momento/list_caches_response_builder.rb
@@ -1,25 +1,0 @@
-require 'grpc'
-require_relative 'generated/controlclient_pb'
-
-module Momento
-  # @private
-  class ListCachesResponseBuilder < ResponseBuilder
-    # Build a Momento::ListCachesResponse from a block of code
-    # which returns a Momento::ControlClient::ListCachesResponse..
-    #
-    # @return [Momento::ListCachesResponse]
-    # @raise [StandardError] when the exception is not recognized.
-    # @raise [TypeError] when the response is not recognized.
-    def from_block
-      response = yield
-    rescue GRPC::BadStatus => e
-      ListCachesResponse::Error.new(
-        exception: e, context: context
-      )
-    else
-      raise TypeError unless response.is_a?(::MomentoProtos::ControlClient::PB__ListCachesResponse)
-
-      return ListCachesResponse::Success.new(grpc_response: response)
-    end
-  end
-end
