+++ lib/momento/response/control/create_cache_response.rb
@@ -1,7 +1,41 @@
-require 'grpc'
-require_relative 'generated/controlclient_pb'
+require_relative '../error'
+require_relative '../../generated/controlclient_pb'
 
 module Momento
+  # A response from creating a cache.
+  class CreateCacheResponse < Response
+    # Does the cache already exist?
+    # @return [Boolean]
+    def already_exists?
+      false
+    end
+
+    # Was the cache created?
+    # @return [Boolean]
+    def success?
+      false
+    end
+
+    # @private
+    class AlreadyExists < CreateCacheResponse
+      def already_exists?
+        true
+      end
+    end
+
+    # @private
+    class Success < CreateCacheResponse
+      def success?
+        true
+      end
+    end
+
+    # @private
+    class Error < CreateCacheResponse
+      include ::Momento::Response::Error
+    end
+  end
+
   # @private
   class CreateCacheResponseBuilder < ResponseBuilder
     # Build a Momento::CreateCacheResponse from a block of code
