+++ lib/momento/response/control/delete_cache_response.rb
@@ -1,7 +1,28 @@
-require 'grpc'
-require_relative 'generated/controlclient_pb'
+require_relative '../error'
+require_relative '../../generated/controlclient_pb'
 
 module Momento
+  # A response from deleting a cache.
+  class DeleteCacheResponse < Response
+    # Was the cache deleted?
+    # @return [Boolean]
+    def success?
+      false
+    end
+
+    # @private
+    class Error < DeleteCacheResponse
+      include ::Momento::Response::Error
+    end
+
+    # @private
+    class Success < DeleteCacheResponse
+      def success?
+        true
+      end
+    end
+  end
+
   # @private
   class DeleteCacheResponseBuilder < ResponseBuilder
     # Build a Momento::DeleteCacheResponse from a block of code
