+++ lib/momento/response/get_response.rb
@@ -1,4 +1,5 @@
-require_relative 'response/error'
+require_relative 'error'
+require_relative '../generated/cacheclient_pb'
 
 module Momento
   # A response containing the value retrieved from a cache.
@@ -69,6 +70,38 @@ module Momento
     # @private
     class Error < GetResponse
       include Momento::Response::Error
+    end
+  end
+
+  # @private
+  class GetResponseBuilder < ResponseBuilder
+    # Build a Momento::GetResponse from a block of code
+    # which returns a Momento::ControlClient::GetResponse.
+    #
+    # @return [Momento::GetResponse]
+    # @raise [StandardError] when the exception is not recognized.
+    # @raise [TypeError] when the response is not recognized.
+    def from_block
+      response = yield
+    rescue *RESCUED_EXCEPTIONS => e
+      GetResponse::Error.new(exception: e, context: context)
+    else
+      from_response(response)
+    end
+
+    private
+
+    def from_response(response)
+      raise TypeError unless response.is_a?(MomentoProtos::CacheClient::PB__GetResponse)
+
+      case response.result
+      when :Hit
+        GetResponse::Hit.new(grpc_response: response)
+      when :Miss
+        GetResponse::Miss.new
+      else
+        raise "Unknown get result: #{response.result}"
+      end
     end
   end
 end
