+++ lib/momento/response/response.rb
@@ -1,25 +1,19 @@
 require 'grpc'
-require_relative 'response/error'
 require_relative 'error'
-require_relative 'error/grpc_details'
-require_relative 'error/transport_details'
-require_relative 'error_builder'
+require_relative '../error'
+require_relative '../error/grpc_details'
+require_relative '../error/transport_details'
+require_relative '../error_builder'
 require_relative 'response_builder'
-require_relative 'create_cache_response'
-require_relative 'create_cache_response_builder'
+require_relative 'control/create_cache_response'
 require_relative 'delete_response'
-require_relative 'delete_response_builder'
-require_relative 'delete_cache_response'
-require_relative 'delete_cache_response_builder'
+require_relative 'control/delete_cache_response'
 require_relative 'get_response'
-require_relative 'get_response_builder'
-require_relative 'list_caches_response'
-require_relative 'list_caches_response_builder'
+require_relative 'control/list_caches_response'
 require_relative 'set_response'
-require_relative 'set_response_builder'
-require_relative 'response/sorted_set/sorted_set_put_element_response'
-require_relative 'response/sorted_set/sorted_set_put_elements_response'
-require_relative 'response/sorted_set/sorted_set_fetch_response'
+require_relative 'sorted_set/sorted_set_put_element_response'
+require_relative 'sorted_set/sorted_set_put_elements_response'
+require_relative 'sorted_set/sorted_set_fetch_response'
 
 module Momento
   # The response from a Momento service request.
