+++ lib/momento/response/sorted_set/sorted_set_put_elements_response.rb
@@ -1,4 +1,3 @@
-require 'grpc'
 require_relative '../error'
 require_relative '../../generated/cacheclient_pb'
