+++ spec/momento/create_cache_response/already_exists_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::CreateCacheResponse::AlreadyExists do
   let(:response) {
