+++ spec/momento/create_cache_response/error_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::CreateCacheResponse::Error do
   let(:response) {
