+++ spec/momento/create_cache_response/success_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::CreateCacheResponse::Success do
   let(:response) {
