+++ spec/momento/delete_cache_response/error_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::DeleteCacheResponse::Error do
   let(:response) {
