+++ spec/momento/delete_response/error_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::DeleteResponse::Error do
   let(:response) {
