+++ spec/momento/delete_response/success_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::DeleteResponse::Success do
   let(:response) {
