+++ spec/momento/get_response/hit_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::GetResponse::Hit do
   let(:grpc_response) {
