+++ spec/momento/get_response/miss_spec.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.describe Momento::GetResponse::Miss do
   let(:response) {
