+++ spec/support/shared_examples/momento/list_caches_response.rb
@@ -1,4 +1,4 @@
-require 'momento/response'
+require 'momento'
 
 RSpec.shared_examples Momento::ListCachesResponse do
   it_behaves_like Momento::Response do
